+++ tests/data/preview_py_310/pep604_union_types_line_breaks.py
@@ -83,6 +83,54 @@ def f(
     ...
 
 
+# don't lose comments
+def f(  # 1
+        looooooooooooooooooooooooooong  # 2
+        :  # 3
+        Loooooooooooooooooooooooooooooooooooooooooooooooong  # 4
+        |  # 5
+        Loooooooooooooooooooooooooooooooooooooooooooooooong  # 6
+        =  # 7
+        3  # 8
+        ):  # 9
+    ...  # 10
+
+
+loooooooooooooooooooooooooooooong: (
+    Loooooooooooooooooooooooooooong  # aoeuaoeuaoeuaoeuaoeuaoeu
+)
+
+
+def foo(
+    loooooooooooooooooooooooooooooooooooooooooooooooooooooooooooong:
+        Loooooooooooooooooooooooooooong,
+): ...
+
+
+def foo(
+    loooooooooooooooong: Tuple[
+        Looooooooooooooooooooooooooooong, Loooooooooooooong
+    ],  # aoeuaoeuaoeu
+): ...
+
+
+def foo(a: loooooooooooooong| loooooooooooooooooooooooooooong| looooooooooong| looooooooooong = 3): ...
+def foo(c: (loooooooooooooong| loooooooooooooooooooooooooooong| looooooooooong| looooooooooong) = 4): ...
+def foo(bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb: bbbbbbbbbbbbbbbbbbbbloooooooooooooong| loooooooooooooooooooooooooooong| looooooooooong| looooooooooong = 5): ...
+def foo(not_too_long_param_name: not_too_long_typehint # but there's a reasonably long comment here
+        ): ...
+def foo(not_too_long_param_name: not_too_long_typehint # but there's a reasonably long comment here
+        | loooooooooooooooooooooooooooong| looooooooooong| looooooooooong = 5): ...
+def foo(a: tuple[int, int,] = (7, 3,)): ...
+def foo(a: (
+    # comment_inside_paren
+            loooooooooooooong
+            # comment2
+            |
+            # comment3
+            loooooooooooooong
+            )): ...
+
 # output
 # This has always worked
 z = (
@@ -158,12 +206,10 @@ def foo(i: (int,)) -> None: ...
 
 def foo(
     i: int,
-    x: (
-        Loooooooooooooooooooooooong
+    x: Loooooooooooooooooooooooong
         | Looooooooooooooooong
         | Looooooooooooooooooooong
-        | Looooooong
-    ),
+        | Looooooong,
     *,
     s: str,
 ) -> None:
@@ -172,16 +218,100 @@ def foo(
 
 @app.get("/path/")
 async def foo(
-    q: str | None = Query(
-        None, title="Some long title", description="Some long description"
-    )
+    q: str
+        | None
+        = Query(None, title="Some long title", description="Some long description")
 ):
     pass
 
 
 def f(
-    max_jobs: int | None = Option(
-        None, help="Maximum number of jobs to launch. And some additional text."
-    ),
+    max_jobs: int
+        | None
+        = Option(
+            None, help="Maximum number of jobs to launch. And some additional text."
+        ),
     another_option: bool = False,
+): ...
+
+
+# don't lose comments
+def f(  # 1
+    looooooooooooooooooooooooooong:  # 2  # 3
+        Loooooooooooooooooooooooooooooooooooooooooooooooong  # 4
+        | Loooooooooooooooooooooooooooooooooooooooooooooooong  # 5  # 6
+        = 3,  # 7  # 8
+):  # 9
+    ...  # 10
+
+
+loooooooooooooooooooooooooooooong: (
+    Loooooooooooooooooooooooooooong  # aoeuaoeuaoeuaoeuaoeuaoeu
+)
+
+
+def foo(
+    loooooooooooooooooooooooooooooooooooooooooooooooooooooooooooong:
+        Loooooooooooooooooooooooooooong,
+): ...
+
+
+def foo(
+    loooooooooooooooong: Tuple[
+        Looooooooooooooooooooooooooooong, Loooooooooooooong
+    ],  # aoeuaoeuaoeu
+): ...
+
+
+def foo(
+    a: loooooooooooooong
+        | loooooooooooooooooooooooooooong
+        | looooooooooong
+        | looooooooooong
+        = 3,
+): ...
+def foo(
+    c: loooooooooooooong
+        | loooooooooooooooooooooooooooong
+        | looooooooooong
+        | looooooooooong
+        = 4,
+): ...
+def foo(
+    bbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbbb:
+        bbbbbbbbbbbbbbbbbbbbloooooooooooooong
+        | loooooooooooooooooooooooooooong
+        | looooooooooong
+        | looooooooooong
+        = 5,
+): ...
+def foo(
+    not_too_long_param_name:
+        not_too_long_typehint,  # but there's a reasonably long comment here
+): ...
+def foo(
+    not_too_long_param_name:
+        not_too_long_typehint  # but there's a reasonably long comment here
+        | loooooooooooooooooooooooooooong
+        | looooooooooong
+        | looooooooooong
+        = 5,
+): ...
+def foo(
+    a: tuple[
+        int,
+        int,
+    ] = (
+        7,
+        3,
+    )
+): ...
+def foo(
+    a:
+        # comment_inside_paren
+        loooooooooooooong
+        # comment2
+        |
+        # comment3
+        loooooooooooooong,
 ): ...